refactor(minimina): dissolve the manager/supervisor duality into plan builders
The managers were runtimes; the supervisor is the runtime now. What was
left of them was plan construction plus a few misplaced path helpers, so
make that the whole job:

- NativeManager -> NativePlanBuilder (native/plan_builder.rs): builds the
  SupervisorPlan and prepares the directories its units write to. The
  vestigial lifecycle is gone: create() was a literal no-op, and
  destroy() only deleted directories that `network delete` removes
  wholesale one line later.
- DockerPlanBuilder (docker/plan_builder.rs) extracted from
  DockerManager, mirroring the native builder. DockerManager shrinks to
  an explicitly-legacy compose shim for the node-level commands, slated
  for deletion once those are ported to supervisor RPCs.
- The supervisor socket path moves onto the plan contract
  (SupervisorPlan::socket_path_in) — builders bake it into the plan, the
  CLI dials it; previously duplicated on both managers.
- Service-log reading moves to DirectoryManager, which owns path layout.
- Both builders import supervisor::plan directly: they depend on the
  plan contract, not on the supervisor runtime.
