+++ src/app/minimina/src/docker/mod.rs
@@ -1,1 +1,2 @@
 pub mod manager;
+pub mod plan_builder;
