+++ src/app/minimina/src/native/mod.rs
@@ -1,4 +1,4 @@
 pub mod keys;
-pub mod manager;
 pub mod mina_locator;
+pub mod plan_builder;
 pub mod port_manager;
